fixed datatype in the header

+++ src/ioAccess.hpp
@@ -11,7 +11,7 @@
   bool ioAccessInitAttachToPwmChannel(uint8_t port, uint8_t channel);
 
   void ioAccessSetDigitalOutput(uint8_t port, bool value);
-  void ioAccessSetPwmUtil(uint8_t channel, uint8_t dutyCycle);
+  void ioAccessSetPwmUtil(uint8_t channel, int dutyCycle);
   bool ioAccessGetDigitalInput(uint8_t port);
   float ioAccessGetAnalogInput(uint8_t port); // scaled to -1 to 1 (or 0-1 if no negative value is possible)
